standalone publisher has it's own cli command
