+++ pype/tools/standalonepublish/__init__.py
@@ -1,8 +1,10 @@
 from .app import (
-    show,
-    cli
+    main,
+    Window
 )
-__all__ = [
-    "show",
-    "cli"
-]
+
+
+__all__ = (
+    "main",
+    "Window"
+)
